+++ MyBlog/src/cn/seu/bingluo/entity/Tag.java
@@ -42,7 +42,7 @@ public class Tag {
 	/**
 	 * @return the blogCount
 	 */
-	public long getBlogCount() {
+	public int getBlogCount() {
 		return blogCount;
 	}
